+++ salt/fileprotos/__init__.py
@@ -1,0 +1,4 @@
+# -*- coding: utf-8 -*-
+'''
+Url Client Module Directory
+'''
